% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model-slurm.R
\name{submit_parallel_model_slurm}
\alias{submit_parallel_model_slurm}
\title{submit a nonmem model to slurm in parallel}
\usage{
submit_parallel_model_slurm(
  .mod,
  partition = "cpu4",
  ncpu = 4,
  overwrite = FALSE,
  slurm_job_template_path = .dh$slurm_job_template_path,
  submission_root = .dh$submission_root,
  bbi_config_path = .dh$bbi_config_path
)
}
\arguments{
\item{.mod}{a bbi nonmem model object}

\item{partition}{name of the partition to submit the model}

\item{ncpu}{number of cpus to run the model against}

\item{overwrite}{whether to overwrite existing model results}

\item{slurm_job_template_path}{path to slurm job template}

\item{submission_root}{directory to track job submission scripts and output}

\item{bbi_config_path}{path to bbi config file}
}
\description{
submit a nonmem model to slurm in parallel
}
\details{
check the partition names using \code{sinfo} from the terminal, provide the partition
name to submit to different partitions (eg general, cpu4, cpu16, etc)
}
